Avoid neo4j crash when indexing long string
When string property byte array length is bigger than 32768, lucene will throw Exception due to LUCENE-5472 improvement, which leads neo4j engine stopped to response. This modification change StringField to TextField, and BinaryDocValuesField instead of SortedSetDocValuesField when longer than 10240 (for UTF8 encoding, the String bytes length are close to 10922=32766/3). It is only a workaround solution.